feat: constant long support
This commit adds support for bigger number of constants in chunk. Before it was limited to one Opcode CONSTANT which took 1 byte argument, resulting in maximum index of 256. Now the chunk will write CONSTANT_LONG opcode if the number of constants will grow bigger than 256, which is 24 bits large, allowing us to store 2^24 constants in a chunk.