feat: parse unary and grouping
This commit adds rest of the logic to parse the unary and grouping expression. It also adds few tests for them, as well as for the Token method to get the binding power. Unit testing in the parser is filled with quite a lot of boilerplate atm. I will probably work on some macro utility to make things shorter / easier to instantiate.