feat: strings
This commit adds support for the strings as a chunk constant. It also adds few trait implementations for the +-/* operators, so we can call them on Values and in return we will get a Result. Addition is possible on numbers and strings. In case of strings they are concatenated, on numbers it's standard math operation. Rest of the math operations are the same and are possible only to perform on numbers. I simplified it a bit with little macro. Also, I fixed a little bug that caused the Text regex to treat text with space inside as an error.