feat: variable declaration and assignment
This commit finishes the work related to variables. The compiler is now able to parse the Expr::Var and Stmt::Var in order to generate bytecode that instantiates and assigns values to variables. Most of the work is done during the generation of the bytecode, so VM doesn't really have to do any lookups, because all of the information about where to get variable from and where assign new value to is calculated and saved in bytecode.