feat: closures
This commit adds support for closures.
It uses the similar approach to the one described in the Crafting Interpreters in the "Closures" chapter or more specifically the one used by the Lua language. You can learn more at https://craftinginterpreters.com/closures.html.
This code is raw, it's not yet refactored and needs to be thoroughly tested, although it works fine in most of the basic use cases, so if there are any bugs, they will be probably spotted in some edge cases.