feat: basic creation of new object instances and parsing of the whole syntax responsible for OOP stuff
Lots of stuff going in here. The parser is already wired up to parse nearly all of the syntax related to OOP, but only bytecode and VM's logic responsible for creation of the object instances is there at the moment. The syntax to create new objects is mostly based on the one seen in Rust, but with minor tweaks.