feat: start of a total refactor
This commit introduces complete refactor of the project structure to separate each part of the language into separate crate and to omit problems and bad architecture design decisions I stumbled upon during previous iteration where I mostly learned new stuff.
This time everything is gonna be built with TDD in mind and I'm starting with a pre-designed idea of the grammar I want to parse, so I can plan everything and not just write random code until it works :D
My plan is to completely parse everything before going on to bytecode generation. The generation is gonna look completely different right now because I believe I made the wrong choice basing the generator on visitor pattern which after some time was really hard to understand and not quite convenient to test.